# ♻️ Refactor: improve iterator performance.