--- a/client/request.go
+++ b/client/request.go
@@ -10,6 +10,7 @@ import (
 	"path/filepath"
 	"reflect"
 	"slices"
+	"sort"
 	"strconv"
 	"sync"
 	"time"
@@ -128,6 +129,24 @@ func (r *Request) Header(key string) []string {
 	return r.header.PeekMultiple(key)
 }
 
+type pair struct {
+	k []string
+	v []string
+}
+
+func (p *pair) Len() int {
+	return len(p.k)
+}
+
+func (p *pair) Swap(i, j int) {
+	p.k[i], p.k[j] = p.k[j], p.k[i]
+	p.v[i], p.v[j] = p.v[j], p.v[i]
+}
+
+func (p *pair) Less(i, j int) bool {
+	return p.k[i] < p.k[j]
+}
+
 // Headers returns an iterator over all headers in the Request.
 // Use maps.Collect() to gather them into a map if needed.
 //
@@ -135,19 +154,24 @@ func (r *Request) Header(key string) []string {
 // Do not store references to returned values; make copies instead.
 func (r *Request) Headers() iter.Seq2[string, []string] {
 	return func(yield func(string, []string) bool) {
-		peekKeys := r.header.PeekKeys()
-		keys := make([][]byte, len(peekKeys))
-		copy(keys, peekKeys) // It is necessary to have immutable byte slice.
+		vals := r.header.Len()
+		p := pair{
+			k: make([]string, 0, vals),
+			v: make([]string, 0, vals),
+		}
+		for k, v := range r.header.All() {
+			p.k = append(p.k, utils.UnsafeString(k))
+			p.v = append(p.v, utils.UnsafeString(v))
+		}
+		sort.Sort(&p)
 
-		for _, key := range keys {
-			vals := r.header.PeekAll(utils.UnsafeString(key))
-			valsStr := make([]string, len(vals))
-			for i, v := range vals {
-				valsStr[i] = utils.UnsafeString(v)
-			}
-
-			if !yield(utils.UnsafeString(key), valsStr) {
-				return
+		j := 0
+		for i := 0; i < vals; i++ {
+			if i == vals-1 || p.k[i] != p.k[i+1] {
+				if !yield(p.k[i], p.v[j:i+1]) {
+					break
+				}
+				j = i
 			}
 		}
 	}
@@ -195,21 +219,24 @@ func (r *Request) Param(key string) []string {
 // Do not store references to returned values; make copies instead.
 func (r *Request) Params() iter.Seq2[string, []string] {
 	return func(yield func(string, []string) bool) {
-		keys := r.params.Keys()
+		vals := r.formData.Len()
+		p := pair{
+			k: make([]string, 0, vals),
+			v: make([]string, 0, vals),
+		}
+		for k, v := range r.formData.All() {
+			p.k = append(p.k, utils.UnsafeString(k))
+			p.v = append(p.v, utils.UnsafeString(v))
+		}
+		sort.Sort(&p)
 
-		for _, key := range keys {
-			if key == "" {
-				continue
-			}
-
-			vals := r.params.PeekMulti(key)
-			valsStr := make([]string, len(vals))
-			for i, v := range vals {
-				valsStr[i] = utils.UnsafeString(v)
-			}
-
-			if !yield(key, valsStr) {
-				return
+		j := 0
+		for i := 0; i < vals; i++ {
+			if i == vals-1 || p.k[i] != p.k[i+1] {
+				if !yield(p.k[i], p.v[j:i+1]) {
+					break
+				}
+				j = i
 			}
 		}
 	}
@@ -298,14 +325,7 @@ func (r *Request) Cookie(key string) string {
 // Cookies returns an iterator over all cookies.
 // Use maps.Collect() to gather them into a map if needed.
 func (r *Request) Cookies() iter.Seq2[string, string] {
-	return func(yield func(string, string) bool) {
-		for k, v := range *r.cookies {
-			res := yield(k, v)
-			if !res {
-				return
-			}
-		}
-	}
+	return r.cookies.All()
 }
 
 // SetCookie sets a single cookie, overriding any previously set value.
@@ -344,13 +364,7 @@ func (r *Request) PathParam(key string) string {
 // PathParams returns an iterator over all path parameters.
 // Use maps.Collect() to gather them into a map if needed.
 func (r *Request) PathParams() iter.Seq2[string, string] {
-	return func(yield func(string, string) bool) {
-		for k, v := range *r.path {
-			if !yield(k, v) {
-				return
-			}
-		}
-	}
+	return r.path.All()
 }
 
 // SetPathParam sets a single path parameter and value, overriding any previously set value.
@@ -440,21 +454,24 @@ func (r *Request) FormData(key string) []string {
 // Do not store references to returned values; make copies instead.
 func (r *Request) AllFormData() iter.Seq2[string, []string] {
 	return func(yield func(string, []string) bool) {
-		keys := r.formData.Keys()
+		vals := r.formData.Len()
+		p := pair{
+			k: make([]string, 0, vals),
+			v: make([]string, 0, vals),
+		}
+		for k, v := range r.formData.All() {
+			p.k = append(p.k, utils.UnsafeString(k))
+			p.v = append(p.v, utils.UnsafeString(v))
+		}
+		sort.Sort(&p)
 
-		for _, key := range keys {
-			if key == "" {
-				continue
-			}
-
-			vals := r.formData.PeekMulti(key)
-			valsStr := make([]string, len(vals))
-			for i, v := range vals {
-				valsStr[i] = utils.UnsafeString(v)
-			}
-
-			if !yield(key, valsStr) {
-				return
+		j := 0
+		for i := 0; i < vals; i++ {
+			if i == vals-1 || p.k[i] != p.k[i+1] {
+				if !yield(p.k[i], p.v[j:i+1]) {
+					break
+				}
+				j = i
 			}
 		}
 	}
@@ -769,20 +786,12 @@ func (c Cookie) DelCookies(key ...string) {
 //
 // The returned key and value should not be retained after the iteration loop.
 func (c Cookie) All() iter.Seq2[string, string] {
-	return func(yield func(string, string) bool) {
-		for k, v := range c {
-			if !yield(k, v) {
-				break
-			}
-		}
-	}
+	return maps.All(c)
 }
 
 // Reset clears the Cookie map.
 func (c Cookie) Reset() {
-	for k := range c {
-		delete(c, k)
-	}
+	clear(c)
 }
 
 // PathParam is a map used to store path parameters.
@@ -825,20 +834,12 @@ func (p PathParam) DelParams(key ...string) {
 //
 // The returned key and value should not be retained after the iteration loop.
 func (p PathParam) All() iter.Seq2[string, string] {
-	return func(yield func(string, string) bool) {
-		for k, v := range p {
-			if !yield(k, v) {
-				break
-			}
-		}
-	}
+	return maps.All(p)
 }
 
 // Reset clears the PathParam map.
 func (p PathParam) Reset() {
-	for k := range p {
-		delete(p, k)
-	}
+	clear(p)
 }
 
 // FormData wraps fasthttp.Args for URL-encoded bodies and form data.