# Description

Reduce iterator time complexity from **O(n<sup>2</sup>)** to **O(n log n)**

+ `Request.Params()`
+ `Request.AllFormData()`

```
goos: darwin
goarch: arm64
pkg: github.com/gofiber/fiber/v3/client
cpu: Apple M4
                        │   old.txt    │           new.txt           │
                        │    sec/op    │   sec/op     vs base        │
_Request_Params-10        135.70n ± 0%   41.01n ± 0%  -69.78% (n=50)
_Request_AllFormData-10   135.60n ± 0%   94.67n ± 0%  -30.18% (n=50)
geomean                    135.6n        62.31n       -54.07%

                        │   old.txt   │          new.txt           │
                        │    B/op     │    B/op     vs base        │
_Request_Params-10        232.00 ± 0%   88.00 ± 0%  -62.07% (n=50)
_Request_AllFormData-10    232.0 ± 0%   184.0 ± 0%  -20.69% (n=50)
geomean                    232.0        127.2       -45.15%

                        │  old.txt   │          new.txt           │
                        │ allocs/op  │ allocs/op   vs base        │
_Request_Params-10        9.000 ± 0%   4.000 ± 0%  -55.56% (n=50)
_Request_AllFormData-10   9.000 ± 0%   6.000 ± 0%  -33.33% (n=50)
geomean                   9.000        4.899       -45.57%
```

Refactor: call existing functions directly

+ `Request.Cookies()` 
+ `Request.PathParams()`
+ `Cookie.All()`
+ `Cookie.Reset()`
+ `PathParam.All()`
+ `PathParam.Reset()`

## Type of change

- [x] Performance improvement (non-breaking change which improves efficiency)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [x] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [x] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [x] Aimed for optimal performance with minimal allocations in the new code.
- [ ] Provided benchmarks for the new code to analyze and improve upon.